cleaned up ZmapDialog
now it is much smarter about working with different types.
Also added ZmapDialog.AddDurationEdit  and a test function

+++ src/bdiff2.m
@@ -97,7 +97,7 @@ classdef bdiff2
                     return
                 end
             else
-                dlg_res.mc_method = double(McMethods.MaxCurvature);
+                dlg_res.mc_method = McMethods.MaxCurvature;
                 dlg_res.fMccorr = fMccorr;
                 dlg_res.doBootstrap = false;
                 dlg_res.nBstSample = nBstSample;
